fix: 调整限流配置解决登录问题
问题：
- 之前的限流配置过于严格（每秒5次，突发10次）
- 开发测试时频繁登录容易触发429错误

解决方案：
- 调整限流参数：每秒10次请求，突发20次
- 仅对认证接口（/register, /login）应用限流
- 使用令牌桶算法基于IP地址限流

改进：
- 正常使用不会触发限流
- 仍能有效防止暴力破解攻击
- 适合开发和生产环境

相关文档：
- docs/RATE_LIMITING.md - 限流配置详细说明
- docs/RATE_LIMITING_FIX.md - 问题修复总结
- scripts/test_rate_limit.sh - 限流功能测试脚本

+++ menu-backend/src/main.rs
@@ -96,7 +96,7 @@ async fn check(config: &AppConfig) -> (PgPool, MultiplexedConnection, Box<Bucket
             std::process::exit(1);
         });
 
-    let row: (i64,) = sqlx::query_as("SELECT 1")
+    let row: (i32,) = sqlx::query_as("SELECT 1")
         .fetch_one(&pool)
         .await
         .expect("数据库查询失败");
